+++ src/main/java/br/com/java/tcc/application/products/ProductServiceImpl.java
@@ -3,13 +3,19 @@ package br.com.java.tcc.application.products;
 import br.com.java.tcc.application.company.CompanyService;
 import br.com.java.tcc.application.company.persistence.CompanyEntity;
 import br.com.java.tcc.application.company.persistence.CompanyRepository;
+import br.com.java.tcc.application.people.persistence.PersonEntity;
 import br.com.java.tcc.application.products.persistence.ProductEntity;
 import br.com.java.tcc.application.products.persistence.ProductRepository;
 import br.com.java.tcc.application.products.resources.ProductRequest;
 import br.com.java.tcc.application.products.resources.ProductResponse;
 import br.com.java.tcc.application.products.util.ProductMapper;
+import br.com.java.tcc.configuration.MessageCodeEnum;
+import br.com.java.tcc.configuration.MessageConfiguration;
+import br.com.java.tcc.exceptions.CustomException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -25,6 +31,9 @@ public class ProductServiceImpl implements ProductService {
     private final ProductMapper productMapper;
 
     private final CompanyService companyService;
+
+    @Autowired
+    MessageConfiguration messageConfiguration;
 
     @Override
     public ProductResponse findById(Long id) {
@@ -50,19 +59,26 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public ProductResponse update (Long id, ProductRequest productDTO){
+        if (!productRepository.existsById(id)) {
+            throw new CustomException(messageConfiguration.getMessageByCode(MessageCodeEnum.REGISTER_NOT_FOUND, "(Produto)"), HttpStatus.NOT_FOUND);
+        }
 
         ProductEntity productEntity = returnProducts(id);
         productMapper.updateProductData(productEntity, productDTO);
-        return productMapper.toProductDTO(productRepository.save(productEntity));
+        ProductEntity updateEntity = productRepository.save(productEntity);
+        return productMapper.toProductDTO(updateEntity);
     }
     @Override
     public String delete(Long id){
+        if (!productRepository.existsById(id)) {
+            throw new CustomException(messageConfiguration.getMessageByCode(MessageCodeEnum.REGISTER_NOT_FOUND, "(Produto)"), HttpStatus.NOT_FOUND);
+        }
         productRepository.deleteById(id);
-        return "Product id: " + id + " deleted";
+        return "Produto id: " + id + " deletado";
     }
 
     public ProductEntity returnProducts(Long id) {
         return productRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Product wasn't found on database"));
+                .orElseThrow(()-> new CustomException(messageConfiguration.getMessageByCode(MessageCodeEnum.REGISTER_NOT_FOUND, "(Produto)"), HttpStatus.NOT_FOUND));
     }
 }
