+++ src/main/java/br/com/java/tcc/application/services/ServicesServiceImpl.java
@@ -8,8 +8,13 @@ import br.com.java.tcc.application.services.persistence.ServicesRepository;
 import br.com.java.tcc.application.services.resources.ServicesRequest;
 import br.com.java.tcc.application.services.resources.ServicesResponse;
 import br.com.java.tcc.application.services.util.ServicesMapper;
+import br.com.java.tcc.configuration.MessageCodeEnum;
+import br.com.java.tcc.configuration.MessageConfiguration;
+import br.com.java.tcc.exceptions.CustomException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,6 +24,9 @@ import java.util.Optional;
 @Primary
 @RequiredArgsConstructor
 public class ServicesServiceImpl implements ServicesService {
+
+    @Autowired
+    MessageConfiguration messageConfiguration;
 
     private final ServicesRepository servicesRepository;
 
@@ -50,6 +58,9 @@ public class ServicesServiceImpl implements ServicesService {
 
     @Override
     public ServicesResponse update(Long id, ServicesRequest servicesDTO){
+        if (!servicesRepository.existsById(id)) {
+            throw new CustomException(messageConfiguration.getMessageByCode(MessageCodeEnum.REGISTER_NOT_FOUND, "(Serviço)"), HttpStatus.NOT_FOUND);
+        }
 
         ServicesEntity servicesEntity = returnServices(id);
         servicesMapper.updateServicesData(servicesEntity,servicesDTO);
@@ -58,12 +69,15 @@ public class ServicesServiceImpl implements ServicesService {
 
     @Override
     public  String delete(Long id){
+        if (!servicesRepository.existsById(id)){
+            throw new CustomException(messageConfiguration.getMessageByCode(MessageCodeEnum.REGISTER_NOT_FOUND, "(Serviço)"), HttpStatus.NOT_FOUND);
+        }
         servicesRepository.deleteById(id);
-        return "Services id: " + id + " deleted";
+        return "Serviço id: " + id + " deletado";
     }
 
     public ServicesEntity returnServices(Long id) {
         return servicesRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Services wasn't found on database"));
+                .orElseThrow(()-> new CustomException(messageConfiguration.getMessageByCode(MessageCodeEnum.REGISTER_NOT_FOUND, "(Serviço)"), HttpStatus.NOT_FOUND));
     }
 }
